+++ include/build_version.h
@@ -8,10 +8,10 @@
 #ifndef BUILD_VERSION_H
 #define BUILD_VERSION_H
 
-#define BUILD_NUMBER 906
-#define BUILD_TIMESTAMP "2025-12-31 13:21:58"
-#define BUILD_DATE "20251231"
-#define GIT_HASH "648de03"
+#define BUILD_NUMBER 918
+#define BUILD_TIMESTAMP "2026-01-01 13:59:20"
+#define BUILD_DATE "20260101"
+#define GIT_HASH "a2168ca"
 #define GIT_BRANCH "main"
 
 #endif // BUILD_VERSION_H